Fix error at starring with operator

+++ src/application/Calculator.java
@@ -1,7 +1,7 @@
 package application;
 
 public class Calculator {
-	public long calculate(long number1, long number2, String operator) {
+	public double calculate(double number1, double number2, String operator) {
         switch (operator) {
             case "+":
                 return number1 + number2;
